+++ libs/email-builder/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/email-builder';
+export * from './lib/types';
 export * from './lib/service';
