+++ libs/email-builder/src/lib/messages/index.ts
@@ -1,0 +1,2 @@
+export * from './new-draft-message';
+export * from './test-message';
