+++ libs/email-builder/src/lib/service.ts
@@ -1,7 +1,7 @@
 import type { SendEmailCommandOutput, SESClient } from '@aws-sdk/client-ses';
 import type { EmailEnvInfo } from '@newsletters-nx/newsletters-data-client';
 import { buildSendEmailCommand } from './build-send-email-command';
-import { buildNewDraftEmail, buildTestEmail } from './email-builder';
+import { buildNewDraftEmail, buildTestEmail } from './messages';
 import type { MessageParams } from './types';
 
 const getMessage = (params: MessageParams, emailEnvInfo: EmailEnvInfo) => {
